+++ integrations/calltaxi/client/client.ts
@@ -1,9 +1,9 @@
 import { createNotification } from "~/utils/notification";
-import { TextureTypes } from "~/types/TextureTypes";
-import { IconTypes } from "~/types/IconTypes";
-import { EventData, Events } from "~/types/Events";
+import { TextureTypes } from "~/types/texture-types";
+import { IconTypes } from "~/types/icon-types";
+import { EventData, Events, SnCommands } from "~/types/events";
 
-emit("chat:addSuggestion", "/calltaxi ", "Contact the taxi services (SnailyCAD).", [
+emit("chat:addSuggestion", `/${SnCommands.CallTaxi}`, "Contact the taxi services.", [
   { name: "description", help: "The description of the call" },
 ]);