+++ integrations/calltow/fxmanifest.lua
@@ -3,7 +3,7 @@ game "gta5"
 
 author "Dev-CasperTheGhost"
 description "/calltow command that connects to SnailyCADv4"
-version "0.0."
+version "0.5.0"
 
 server_scripts {"server/server.js"}
 client_scripts {"client/client.js"}